rename: categories 분리 및 완결 작품 data 형식 변경
