+++ src/nav-category.js
@@ -1,21 +1,8 @@
-import BookCategory from "../screens/Book/BookCategory.js";
-import BroadcastCategory from "../screens/Broadcast/BroadcastCategory.js";
-import HomeCategory from "../screens/Home/HomeCategory.js";
-import MovieCategory from "../screens/Movie/MovieCategory.js";
-import WebnovelCategory from "../screens/Webnovel/WebnovelCategory.js";
-import WebtoonCategory from "../screens/Webtoon/WebtoonCategory.js";
 import { renderCategoryContents } from "./core.js";
+import categories from "../categories.js";
+
 const headerNav = document.querySelector(".header__nav");
 const nav = document.querySelector(".main__navGenre");
-
-const categories = {
-  home: HomeCategory,
-  webtoon: WebtoonCategory,
-  webnovel: WebnovelCategory,
-  movie: MovieCategory,
-  broadcast: BroadcastCategory,
-  book: BookCategory,
-};
 
 const handleNavCategory = (event) => {
   const target = event.target.closest(".header__nav-item");
